+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - Makefile: `makefile-code`, `makefileblock`, `makefileblock0`
   - Markdown: `markdown-code`, `markdownblock`, `markdownblock0`
   - Objective-C: `objc-code`, `objcblock`, `objcblock0`
+  - plist: `plist-code`, `plistblock`, `plistblock0`
   - Racket: `racket-code`, `racketblock`, `racketblock0`
   - Rhombus: `rhombus-code`, `rhombusblock`, `rhombusblock0`
   - Swift: `swift-code`, `swiftblock`, `swiftblock0`